requirements.txt
langchain
langgraph
langchain-core
typing-extensions
rich
fastapi
uvicorn
sse-starlette
python-dotenv
aiofiles
langchain-openai
